feat(bt): implement Bluetooth RFCOMM connection, bufferbloat mitigation, and one-way input lag telemetry

+++ app/src/main/java/com/sanket/tools/nexpad/network/AdbBridgeConnection.kt
@@ -80,7 +80,7 @@ class AdbBridgeConnection(private val context: Context) : IGamepadConnection {
     override var onConnectionStateChanged: ((Boolean) -> Unit)? = null
     override var onStatusChanged: ((String) -> Unit)? = null
     override var onDiagnosticLog: ((String) -> Unit)? = null
-    override var onNetworkPerformanceUpdated: ((latencyMs: Long, jitterMs: Long, packetLoss: Float) -> Unit)? = null
+    override var onNetworkPerformanceUpdated: ((latencyMs: Long, jitterMs: Float, packetLoss: Float) -> Unit)? = null
 
     override suspend fun connect(address: String, port: Int) {
         withContext(Dispatchers.IO) {
@@ -282,7 +282,7 @@ class AdbBridgeConnection(private val context: Context) : IGamepadConnection {
             val lossPctFloat = (packetLossByte and 0xFF) / 255f
 
             callbackScope?.launch {
-                onNetworkPerformanceUpdated?.invoke(avgRtt.toLong(), jitterMs.toLong(), lossPctFloat)
+                onNetworkPerformanceUpdated?.invoke(avgRtt.toLong(), jitterMs.toFloat(), lossPctFloat)
             }
         }
     }
